Ensure attachment MIME type matches that in ACP settings
This ensures that the new plugin setting 'Video (or other) MIME types'
reliably matches the MIME type of attachments as stipulated in the ACP's
Attachment Types settings.

It is necessary because MyBB core code stores to the attachments table
the MIME type which the attachment was uploaded as, which doesn't
necessarily match the MIME type for files with that extension in
Attachment Types.